app/pyproject.toml
[tool.poetry]
name = "agni"
version = "0.1.0"
description = ""
authors = ["Yusuke Shimizu <44394399+ushmz@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.6.7"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
alembic = "^1.5.8"
mysql-connector-python = "^8.0.25"
sqlalchemy-stubs = "^0.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.1"
black = "^21.4b2"
icecream = "^2.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
